--- Bogdan/7 lab/ConsoleApp 7/ConsoleApp 7.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29920.165
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Lababa7AllahuAkbar", "Lababa7AllahuAkbar\Lababa7AllahuAkbar.vcxproj", "{8433B516-D87B-4C74-83F2-2855C4FF0D6F}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8433B516-D87B-4C74-83F2-2855C4FF0D6F}.Debug|x64.ActiveCfg = Debug|x64
-		{8433B516-D87B-4C74-83F2-2855C4FF0D6F}.Debug|x64.Build.0 = Debug|x64
-		{8433B516-D87B-4C74-83F2-2855C4FF0D6F}.Debug|x86.ActiveCfg = Debug|Win32
-		{8433B516-D87B-4C74-83F2-2855C4FF0D6F}.Debug|x86.Build.0 = Debug|Win32
-		{8433B516-D87B-4C74-83F2-2855C4FF0D6F}.Release|x64.ActiveCfg = Release|x64
-		{8433B516-D87B-4C74-83F2-2855C4FF0D6F}.Release|x64.Build.0 = Release|x64
-		{8433B516-D87B-4C74-83F2-2855C4FF0D6F}.Release|x86.ActiveCfg = Release|Win32
-		{8433B516-D87B-4C74-83F2-2855C4FF0D6F}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {62529DD9-C2CB-4518-8915-F067CA7EFF41}
-	EndGlobalSection
-EndGlobal
