--- Bogdan/7 lab/ConsoleApp 7/ConsoleApp 7/Debug/Console App 7.log
@@ -1,14 +1,0 @@
-﻿  HotelsMain.cpp
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(32,10): warning C4473: "scanf_s": передано недостаточно аргументов для строки формата
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(32,10): message : заполнители и их параметры ожидают переменное число аргументов в количестве 2, но было предоставлено 1
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(32,10): message : отсутствующий вариативный аргумент "2" требуется для строки форматирования "%c"
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(32,10): message : этот аргумент используется в качестве размера буфера
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(75,16): warning C4018: <: несоответствие типов со знаком и без знака
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(88,20): warning C4018: <: несоответствие типов со знаком и без знака
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(100,20): warning C4018: <: несоответствие типов со знаком и без знака
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(120,17): warning C4018: <: несоответствие типов со знаком и без знака
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(142,11): warning C4477: "printf": в строке форматирования "%d" требуется аргумент типа "int", но вариативный аргумент "3" имеет тип "char *"
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(160,20): warning C4018: <: несоответствие типов со знаком и без знака
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(254): warning C4715: StringParse: значение возвращается не при всех путях выполнения
-D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Lababa7AllahuAkbar\HotelsMain.cpp(252): warning C4172: возвращение адреса локальной или временной переменной : final_word
-  Lababa7AllahuAkbar.vcxproj -> D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\Debug\Lababa7AllahuAkbar.exe
