--- Bogdan/7 lab/ConsoleApp 7/ConsoleApp 7/Debug/Console App 7A.8433B516.tlog/ConsoleApp7.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.25.28610:TargetPlatformVersion=10.0.18362.0:
-Debug|Win32|D:\Laby\C\C-Lab-Number7-Var9-master\Lababa7AllahuAkbar\|
